+++ src/Model/Validation/PostValidator.php
@@ -80,8 +80,8 @@ class PostValidator extends AppValidator {
 			if(!empty($tag['id']))
 				continue;
 			
-			//Checks if the tag has between 3 and 20 chars
-			if(empty($tag['tag']) || strlen($tag['tag']) < 3 || strlen($tag['tag'] > 20))
+			//Checks if the tag has between 3 and 30 chars
+			if(empty($tag['tag']) || strlen($tag['tag']) < 3 || strlen($tag['tag'] > 30))
 				return FALSE;
 		}