+++ src/Controller/Admin/PostsCategoriesController.php
@@ -104,7 +104,6 @@ class PostsCategoriesController extends AppController {
     /**
      * Edits posts category
      * @param string $id Posts category ID
-     * @throws \Cake\Network\Exception\NotFoundException
      */
     public function edit($id = NULL)  {
         $category = $this->PostsCategories->get($id);
@@ -125,7 +124,6 @@ class PostsCategoriesController extends AppController {
     /**
      * Deletes posts category
      * @param string $id Posts category ID
-     * @throws \Cake\Network\Exception\NotFoundException
      */
     public function delete($id = NULL) {
         $this->request->allowMethod(['post', 'delete']);