+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 2.1 branch
 ### 2.1.7
 * an exception is now properly thrown when a record is not found;
+* tags can be 30 characters long;
 * fixed bug. Now managers can access "system" menu.
 
 ### 2.1.6