+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # 2.x branch
 ## 2.1 branch
+### 2.1.7
+* an exception is now properly thrown when a record is not found;
+* tags can be 30 characters long;
+* fixed bug. Now managers can access "system" menu.
+
 ### 2.1.6
 * the User shell can now list users and user groups;
 * added "latest photos" widget;